bio-lab/pyproject.toml
[tool.poetry]
name = "bio-lab"
version = "0.0.1"
description = "Composition, Validation, and Execution tool for composite simulations relating to Systems Biology"
authors = ["Alex Patrie <alexanderpatrie@gmail.com>"]
readme = "README.md"
packages = [{include = "biolab"}]

# TODO: Remove duplicate deps and just import/install biosimulator-processes
[tool.poetry.dependencies]
python = ">=3.10"
zarr = "*"
termcolor = "*"
jupyterlab = "*"
bigraph-schema = "*"
process-bigraph = "^0.0.17"
bigraph-viz = "*"
bigraph-builder = "*"
numpy = "*"
copasi-basico = "*"
cobra = "*"
tellurium = "*"
amici = "*"
# smoldyn = {version = ">=2.72", optional = true}
python-libsbml = "^5.20.2"
docker = ">=7.0"
python-libnuml = "^1.1.6"
notebook = "^7.1.1"
toml = "*"
fastapi = "^0.109.0"  # "*"
pyyaml = "*"
pydantic = "*"
h5py = "*"
pydantic-settings = "*"
httpx = "*"
tensorstore = "*"
aiofiles = "*"
aiobotocore = "*"
uvicorn = {extras = ["standard"], version = "*"}
seaborn = "^0.13.2"
beautifulsoup4 = "4.12.3"
transformers = "^4.41.0"
openai = "^1.30.1"

[tool.poetry.extras]
smoldyn = ["smoldyn"]
# simulariumio = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
mypy = ">=1.8.0"
flake8 = "*"
twine = "*"

[tool.pytest.ini_options]
testpaths = ["bio_lab"]
addopts = "--ignore=setup.py"
python_files = "*.py"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
